Registration: Fix PHP types if relative date/time records when processing codes
See: https://mantis.ilias.de/view.php?id=44027
(cherry picked from commit 7cdc128f39a70c87f6382b48056ff7079fff2cc2)

+++ components/ILIAS/Registration/classes/class.ilRegistrationSettingsGUI.php
@@ -951,7 +951,7 @@ class ilRegistrationSettingsGUI
                         $date = serialize([
                             "d" => $date["dd"],
                             "m" => $date["MM"] % 12,
-                            "y" => floor($date["MM"] / 12)
+                            "y" => (int) floor($date["MM"] / 12)
                         ]);
                     }
                     break;
